bot combinado/requirements.txt
    #Python 3.10.11
# --- Google Cloud Web Risk ---
google-cloud-webrisk==1.19.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
proto-plus==1.26.1
protobuf==6.33.0
rsa==4.9.1
cachetools==6.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# --- Telegram Bot + Groq + utilidades ---
pyTelegramBotAPI==4.29.1
telebot==0.0.5
groq==0.31.1
python-dotenv==1.1.1
pydantic==2.11.9
pydantic_core==2.33.2
typing-inspection==0.4.1
anyio==4.10.0
httpx==0.28.1
httpcore==1.0.9
sniffio==1.3.1
h11==0.16.0
distro==1.9.0
annotated-types==0.7.0
exceptiongroup==1.3.0

# --- Requests stack (mantén solo una versión coherente) ---
requests==2.32.5
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11
certifi==2025.8.3

# --- Typing support ---
typing_extensions==4.15.0